Updated Main() structure to allow for more descriptive command error messages.
Added more command error messages and updated old ones.

Added behavior to check if the user wants to overwrite an excel file with the same name as the output file specified for the parse command or cancel the operation.

Added behavior to check if the user wants to overwrite a worksheet in the specified excel file for the analyze command or cancel the operation.

Added startup message.

Added some TODOs, see above Main().

Fixed error where if no file extension accompanied text file to be parsed a .xlsx extension was added.

Moved worksheet generation from Main() to DataLoader.Load() to keep Main()'s functionality to being a menu.

Capitalized a comment in AnalyticsLoader...