Added input cleaning to parse command in Program.
Added handling for parse attempts that result in zero samples to DataLoader.

Added handling for Calibration, Quality Control, or Certified Value Samples that exist in check standards but not in the text file being parsed to DataLoader.

Added Utils class with Levenshtein Distance calculation for testing sample name similarity recognition. It isn't ass good as just using contains it seems.

Changed this parser to parse Elements with no data as Double.NaN to DataLoaderParser (this branches version).